refactor: moved extensions from wallet_chain into the utils
