+++ lib/service/confirmation_dialogs/transaction_confirmed_dialog.dart
@@ -1,5 +1,6 @@
 import 'package:ax_dapp/service/controller/controller.dart';
 import 'package:ax_dapp/service/custom_styles.dart';
+import 'package:ax_dapp/util/chain_localized_name.dart';
 import 'package:ax_dapp/wallet/wallet.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -70,7 +71,7 @@ class LinkButtons extends StatelessWidget {
           const LinkButton(
             pageName: 'Farm',
             page: 'farm',
-          )
+          ),
       ],
     );
   }
@@ -139,7 +140,7 @@ class TransactionConfirmed extends StatelessWidget {
                           ),
                           onPressed: () => Navigator.pop(context),
                         ),
-                      )
+                      ),
                     ],
                   ),
                 ),
@@ -187,7 +188,7 @@ class TransactionConfirmed extends StatelessWidget {
                   isTrade: isTrade,
                   isPool: isPool,
                   isFarm: isFarm,
-                )
+                ),
               ],
             ),
           ),
