+++ lib/util/chain_localized_image.dart
@@ -1,23 +1,21 @@
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
 
-extension EthereumChainLocalizationX on EthereumChain {
-  String get localizedName {
+extension EthereumChainLocalizationImageX on EthereumChain {
+  String get localizedImage {
     switch (this) {
       case EthereumChain.none:
       case EthereumChain.unsupported:
         throw UnsupportedError('$this');
       case EthereumChain.polygonMainnet:
-        return 'Matic Network';
-      case EthereumChain.optimism:
-        return 'Optimism';
-      case EthereumChain.arbitriumOne:
-        return 'Arbitrium One';
+        return 'assets/images/Polygon_Small.png';
       case EthereumChain.goerliTestNet:
-        return 'Goerli Test Network';
+        return 'assets/images/geth.png';
+      case EthereumChain.optimism:
+      case EthereumChain.arbitriumOne:
       case EthereumChain.sxMainnet:
-        return 'SX Network';
+        return 'assets/images/SX_Small.png';
       case EthereumChain.sxTestnet:
-        return 'SX Test Network';
+        return 'assets/images/SX_Small.png';
     }
   }
 }
